fix(warnings): resolve three static analysis warnings
- connectivity.hpp: cast Response to uint16_t to match %hu format specifier
- hadiscovery_spec: NOLINT for read(uint8_t* buf) override that writes to buf
- hadiscovery_spec: split combined uint32_t declaration into two statements

+++ lib/connectivity/src/connectivity.hpp
@@ -232,7 +232,7 @@ public:
   [[nodiscard]] virtual bool sendResponse(Response response, uint16_t command = 0U, uint32_t errCode = 0U) {
     char responseBuffer[responseBufferSize] = { '\0' };
     const int32_t responseBufferActualSize = snprintf_P(responseBuffer, sizeof(responseBuffer),
-      PSTR(R"({"type":%hu,"cmd":%hu,"err":%u})"), static_cast<uint8_t>(response), command, errCode);
+      PSTR(R"({"type":%hu,"cmd":%hu,"err":%u})"), static_cast<uint16_t>(response), command, errCode);
     const bool responseBufferValid = ((responseBufferActualSize >= 0) &&
       (responseBufferActualSize < static_cast<int32_t>(sizeof(responseBuffer))));
     if(!responseBufferValid) { return false; }
